requirements_backup.txt
# Athena平台依赖快照
# 生成时间: 2026-04-19 21:25:00
# 分支: feature/unified-tool-registry
# 用途: 统一工具注册表重构前的依赖备份

# ===== 核心依赖 =====
# Web框架
fastapi>=0.136.0
uvicorn[standard]>=0.44.0
python-multipart>=0.0.26

# 数据验证
pydantic>=2.0
pydantic-settings>=2.0.0

# 数据库
asyncpg>=0.31.0
psycopg2-binary>=2.9.11
psycopg>=3.3.3
redis>=7.4.0

# 认证安全
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# 异步I/O
aiofiles>=23.2.1
aiohttp>=3.9.0
httpx>=0.28.0

# 数据处理
numpy>=2.4.4
pandas>=3.0.0
scikit-learn>=1.3.0

# 向量数据库
qdrant-client>=1.7.0

# 图数据库
neo4j>=6.1.0

# NLP/文本处理
jieba>=0.42.1

# 任务队列
celery>=5.3.0

# 图像处理
pillow>=12.2.0

# 日志
loguru>=0.7.0

# 工具库
click>=8.1.0
python-dotenv>=1.0.0
tenacity>=9.1.0
requests>=2.33.1

# 监控
prometheus-client>=0.24.1
psutil>=7.2.2

# ===== 开发依赖 =====
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0
pytest-xdist>=3.5.0
pytest-timeout>=2.2.0
pytest-benchmark>=4.0.0

# 代码质量
black>=23.12.0
ruff>=0.1.0
mypy>=1.7.0
pre-commit>=3.6.0

# ===== 注意事项 =====
# 1. Python版本要求: ^3.11
# 2. 部分依赖可能有额外选项（如uvicorn[standard]）
# 3. 如需完整依赖列表，请使用: poetry export -f requirements.txt
# 4. 此文件仅用于快速回滚时的参考，不建议直接用于安装
